% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC.R
\name{DC}
\alias{DC}
\title{Generates text output displaying your primary source of food, given your 15N and 13C inputs}
\description{
Prints 'Generates text output displaying your primary source of food, given your 15N and 13C inputs'.
}

